ml_core/requirements.txt
# ML Core Requirements
# Optimized for CPU execution on 8GB RAM systems

# Core ML Libraries
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.0.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Vector Search
faiss-cpu>=1.7.4
# Note: use faiss-cpu for CPU-only systems

# PDF Processing
pdfminer.six>=20221105
PyMuPDF>=1.23.0  # fitz for PDF image extraction
pytesseract>=0.3.10
Pillow>=10.0.0

# OCR (optional but recommended)
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Utilities
numpy>=1.24.0
tiktoken>=0.5.0  # For accurate token counting

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.5.0  # Type checking
